........
  r23484 | george.moellenbrock | 2013-04-01 09:26:45 -0600 (Mon, 01 Apr 2013) | 13 lines
  
  
         New Development: No
              JIRA Issue: CAS-3014
          Ready for Test: Yes
       Interface Changes: VisCal.h
  What Interface Changed: (some private data changed from Int to Int64)
           Test Programs: List test programs
    Put in Release Notes: No
               Module(s): synthesis
             Description: Solve and apply counters now Int64s, now that
  			the python binding can cope with them (r23480).
........


git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/release-4_1_0@23486 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/synthesis/MeasurementComponents/VisCal.h
@@ -349,7 +349,7 @@ private:
   PtrBlock<Matrix<Float>*> currWtScale_;  // [nSpw](nPar,nElm)
 
   // Flag counting
-  Int ndataIn_, nflagIn_, nflagOut_;
+  Int64 ndataIn_, nflagIn_, nflagOut_;
 
   // Print level
   Int prtlev_;
